index.html & favs.json V1 Files
These are the two main files of this assignment. I took the pagination
code fom Marina and the JQUERY Mobile layout from Mauricio and mixed
together, adding some improvements, such as the automated scroll-based
pagination and fixed bugs that were showing up when we added more tweets
to the document.